Avoid extra work when input nodes have stable output paths.
In most cases, the output paths are actually stable. This changes
things around a bit to make this compatibility layer a tad bit more
"pay as you go", by only doing the extra work of clearing and
re-symlinking the `inputPaths` when it is needed.

This cuts down on at least 4 filesystem operations per node:

1. `fs.lstatSync` done by `rimraf.sync` to determine how to remove prior inputPath
2. `fs.unlinkSync` done by `rimraf.sync` to remove old inputPath
3. `fs.lstatSync` done by `symlink-or-copy` to dereference source symlinks
4. `fs.symlinkSync` done by `symlink-or-copy`

Prior to the changes in this PR the above 4 steps would be done on **every** tree
in the entire system for every build and rebuild. After these changes, the above
steps are only executed for initial builds and when the input paths actually need
to change.