Avoid forcing eager calculation of error stacks
Previously, we would force eager calculation of error stack traces even
if we would never need them. This tweaks things slightly, to avoid eager
computation of the `.stack` property.

After this change (and a similar one over in `broccoli-source`) we will
only calculate stack traces when they are needed (e.g. when there is
a build error and we want to display the instantiation stack).